pyproject.toml
[project]
name = "fastapi-app"
version = "0.1.0"
description = "FastAPI Spring Boot analogue"
authors = [
    {name = "lolowq",email = "ne.shatrov@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.116.2,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "alembic (>=1.16.5,<2.0.0)",
    "jinja2 (>=3.1.6,<4.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "pydantic (>=2.11.9,<3.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=8.4.2,<9.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "pytest-cov (>=7.0.0,<8.0.0)"
]
